## Performance Improvement Analysis

### Total Execution Time
- **Before Improvement**: 82.30 seconds
- **After Improvement**: 60.25 seconds
- **Total Time Improvement**: 22.05 seconds (26.8% reduction)

### Performance Metrics Breakdown

| Metric | Before Improvement | After Improvement | Improvement |
|--------|---------------------|-------------------|-------------|
| Total Time | 82.30 seconds | 60.25 seconds | 22.05 seconds ↓ |
| User Time | 25.18 seconds | 25.07 seconds | 0.11 seconds ↓ |
| System Time | 49.69 seconds | 51.16 seconds | -1.47 seconds ↑ |

Signed-off-by: Vipul Gupta (@vipulgupta2048) <vipulgupta2048@gmail.com>